+++ ots/source/luascript.cpp
@@ -78,13 +78,13 @@ int32_t LuaScript::OpenFile(const char *filename)
     SILVER_DMG_LO = atof(getGlobalStringField("silverwanddmg", 3, "0.24").c_str());
     SILVER_DMG_HI = atof(getGlobalStringField("silverwanddmg", 4, "0.55").c_str());
 
-    MANAS_MIN_LVL = atof(getGlobalStringField("manasMinRate", 1, "1.0").c_str());
-    MANAS_MIN_MLVL = atof(getGlobalStringField("manasMinRate", 2, "1.0").c_str());
-    MANAS_MIN_LO = atof(getGlobalStringField("manasMinRate", 3, "0.24").c_str());
+    MANAS_MIN_LVL = (int64_t)atof(getGlobalStringField("manasMinRate", 1, "1.0").c_str());
+    MANAS_MIN_MLVL = (int64_t)atof(getGlobalStringField("manasMinRate", 2, "1.0").c_str());
+    MANAS_MIN_LO = (int64_t)atof(getGlobalStringField("manasMinRate", 3, "0.24").c_str());
 
-    MANAS_MAX_LVL = atof(getGlobalStringField("manasMaxRate", 1, "1.0").c_str());
-    MANAS_MAX_MLVL = atof(getGlobalStringField("manasMaxRate", 2, "1.0").c_str());
-    MANAS_MAX_HI = atof(getGlobalStringField("manasMaxRate", 3, "1.55").c_str());
+    MANAS_MAX_LVL = (int64_t)atof(getGlobalStringField("manasMaxRate", 1, "1.0").c_str());
+    MANAS_MAX_MLVL = (int64_t)atof(getGlobalStringField("manasMaxRate", 2, "1.0").c_str());
+    MANAS_MAX_HI = (int64_t)atof(getGlobalStringField("manasMaxRate", 3, "1.55").c_str());
 
     ACCESS_PROTECT = getGlobalNumber("accessprotect",3);
     ACCESS_REMOTE = getGlobalNumber("accessremote",3);
@@ -182,17 +182,17 @@ int32_t LuaScript::OpenFile(const char *filename)
     CAP_GAIN[VOCATION_PALADIN] = atoi(getGlobalStringField("capgain", VOCATION_PALADIN + 1, "20").c_str());
     CAP_GAIN[VOCATION_KNIGHT] = atoi(getGlobalStringField("capgain", VOCATION_KNIGHT + 1, "25").c_str());
 
-    MANA_GAIN[VOCATION_NONE] = atoi(getGlobalStringField("managain", VOCATION_NONE + 1, "5").c_str());
-    MANA_GAIN[VOCATION_SORCERER] = atoi(getGlobalStringField("managain", VOCATION_SORCERER + 1, "30").c_str());
-    MANA_GAIN[VOCATION_DRUID] = atoi(getGlobalStringField("managain", VOCATION_DRUID + 1, "30").c_str());
-    MANA_GAIN[VOCATION_PALADIN] = atoi(getGlobalStringField("managain", VOCATION_PALADIN + 1, "15").c_str());
-    MANA_GAIN[VOCATION_KNIGHT] = atoi(getGlobalStringField("managain", VOCATION_KNIGHT + 1, "5").c_str());
+    MANA_GAIN[VOCATION_NONE] = atoll(getGlobalStringField("managain", VOCATION_NONE + 1, "5").c_str());
+    MANA_GAIN[VOCATION_SORCERER] = atoll(getGlobalStringField("managain", VOCATION_SORCERER + 1, "30").c_str());
+    MANA_GAIN[VOCATION_DRUID] = atoll(getGlobalStringField("managain", VOCATION_DRUID + 1, "30").c_str());
+    MANA_GAIN[VOCATION_PALADIN] = atoll(getGlobalStringField("managain", VOCATION_PALADIN + 1, "15").c_str());
+    MANA_GAIN[VOCATION_KNIGHT] = atoll(getGlobalStringField("managain", VOCATION_KNIGHT + 1, "5").c_str());
 
-    HP_GAIN[VOCATION_NONE] = atoi(getGlobalStringField("hpgain", VOCATION_NONE + 1, "5").c_str());
-    HP_GAIN[VOCATION_SORCERER] = atoi(getGlobalStringField("hpgain", VOCATION_SORCERER + 1, "5").c_str());
-    HP_GAIN[VOCATION_DRUID] = atoi(getGlobalStringField("hpgain", VOCATION_DRUID + 1, "5").c_str());
-    HP_GAIN[VOCATION_PALADIN] = atoi(getGlobalStringField("hpgain", VOCATION_PALADIN + 1, "10").c_str());
-    HP_GAIN[VOCATION_KNIGHT] = atoi(getGlobalStringField("hpgain", VOCATION_KNIGHT + 1, "15").c_str());
+    HP_GAIN[VOCATION_NONE] = atoll(getGlobalStringField("hpgain", VOCATION_NONE + 1, "5").c_str());
+    HP_GAIN[VOCATION_SORCERER] = atoll(getGlobalStringField("hpgain", VOCATION_SORCERER + 1, "5").c_str());
+    HP_GAIN[VOCATION_DRUID] = atoll(getGlobalStringField("hpgain", VOCATION_DRUID + 1, "5").c_str());
+    HP_GAIN[VOCATION_PALADIN] = atoll(getGlobalStringField("hpgain", VOCATION_PALADIN + 1, "10").c_str());
+    HP_GAIN[VOCATION_KNIGHT] = atoll(getGlobalStringField("hpgain", VOCATION_KNIGHT + 1, "15").c_str());
 
     WEAPON_MUL[VOCATION_NONE] = atoi(getGlobalStringField("weaponmul", VOCATION_NONE+1, "1").c_str());
     WEAPON_MUL[VOCATION_SORCERER] = atoi(getGlobalStringField("weaponmul", VOCATION_SORCERER+1, "1").c_str());
@@ -236,7 +236,7 @@ int32_t LuaScript::OpenFile(const char *filename)
     PROMOTED_VOCATIONS[VOCATION_DRUID] = getGlobalStringField("promoted_vocations",VOCATION_DRUID);
 
     DIE_PERCENT_EXP = atoi(getGlobalStringField("diepercent",1,"7").c_str());
-    DIE_PERCENT_MANA = atoi(getGlobalStringField("diepercent",2,"7").c_str());
+    DIE_PERCENT_MANA = atoll(getGlobalStringField("diepercent",2,"7").c_str());
     DIE_PERCENT_SKILL = atoi(getGlobalStringField("diepercent",3,"7").c_str());
     DIE_PERCENT_EQ = atoi(getGlobalStringField("diepercent",4,"7").c_str());
     DIE_PERCENT_BP = atoi(getGlobalStringField("diepercent",5,"100").c_str());
@@ -244,17 +244,17 @@ int32_t LuaScript::OpenFile(const char *filename)
 
     DIE_PERCENT_SL = getGlobalDouble("diepercent_sl",0.5);
 
-    MANA_SNAKEBITE = atoi(getGlobalStringField("rodmana", 1, "2").c_str());
-    MANA_MOONLIGHT = atoi(getGlobalStringField("rodmana", 2, "3").c_str());
-    MANA_VOLCANIC = atoi(getGlobalStringField("rodmana", 3, "5").c_str());
-    MANA_QUAGMIRE = atoi(getGlobalStringField("rodmana", 4, "8").c_str());
-    MANA_TEMPEST = atoi(getGlobalStringField("rodmana", 5, "13").c_str());
+    MANA_SNAKEBITE = atoll(getGlobalStringField("rodmana", 1, "2").c_str());
+    MANA_MOONLIGHT = atoll(getGlobalStringField("rodmana", 2, "3").c_str());
+    MANA_VOLCANIC = atoll(getGlobalStringField("rodmana", 3, "5").c_str());
+    MANA_QUAGMIRE = atoll(getGlobalStringField("rodmana", 4, "8").c_str());
+    MANA_TEMPEST = atoll(getGlobalStringField("rodmana", 5, "13").c_str());
 
-    MANA_VORTEX = atoi(getGlobalStringField("wandmana", 1, "2").c_str());
-    MANA_DRAGONBREATH = atoi(getGlobalStringField("wandmana", 2, "3").c_str());
-    MANA_PLAGUE = atoi(getGlobalStringField("wandmana", 3, "5").c_str());
-    MANA_COSMIC = atoi(getGlobalStringField("wandmana", 4, "8").c_str());
-    MANA_INFERNO = atoi(getGlobalStringField("wandmana", 5, "13").c_str());
+    MANA_VORTEX = atoll(getGlobalStringField("wandmana", 1, "2").c_str());
+    MANA_DRAGONBREATH = atoll(getGlobalStringField("wandmana", 2, "3").c_str());
+    MANA_PLAGUE = atoll(getGlobalStringField("wandmana", 3, "5").c_str());
+    MANA_COSMIC = atoll(getGlobalStringField("wandmana", 4, "8").c_str());
+    MANA_INFERNO = atoll(getGlobalStringField("wandmana", 5, "13").c_str());
 
     RANGE_SNAKEBITE = atoi(getGlobalStringField("rodrange", 1, "4").c_str());
     RANGE_MOONLIGHT = atoi(getGlobalStringField("rodrange", 2, "3").c_str());
